fix: static export
Fixes static exporting of routes specified in the root configuration
file. Correct URL instance is now used in the rendering context.

Static export now also renders with the `Accept` header set to `text/x-component`,
and so exporting will work not just on initial loading of pages, but also on
client-side navigation to the exported routes.

Static file handler now works for both HTML and RSC exports.

Fixes minor issue with filename formatting in the reporting of the
static export.