fix: dev hmr
Fixes HMR when rendering without any client components. When no client
components used to render the HTML response the renderer will include
the entry module for HMR, but it will not start hydration on the client.

In a production build when no client components were used rendering will
not include any script tags and so there will be no hydration on the
client.