feat: development server commands and search docs (#65)
This PR implements commands and integrated documentation search when
running the development server. The developer can start typing in the
terminal, and the server auto-suggests commands and search results. When
the search term is more than 2 characters it will search the
documentation using Algolia search and selecting a search result will
open it in the default browser.

#### Available commands

- Restart the development server 🔄
- Reload the application 🔥
- Print the server URLs 🔗
- Clear the console 🧹
- Open application in the default browser 🌐
- Quit the development server 🚫

_This PR also decorates core development server messages with emojis to
have more fun while using `@lazarv/react-server`!_