feat: server functions return RSC payload (#93)
Server functions now return result in RSC payload. This helps with
supporting multiple types of results (including binary formats) and also
enables the new `reload()` API which enables single-roundtrip mutation
and page (or outlet) refresh by combining the server function result
with the rendered component in a single RSC payload.

Includes tests for multiple server function result types and
documentation update to include `reload()`.

This PR addresses an issue with server functions without a return value
(`Promise<void>`) and enhances the framework to be able to return a
rendered component along with the server function return value discussed
in #85.