fix: static generated RSC x-component and response type selection (#99)
This PR fixes issues around static generated RSC x-component files and
how response type selection works to send an RSC payload. The response
type selection needs to be on par with the static generation to maintain
the same solution for both static and dynamic rendering.

Changes request payload to select RSC rendering. Instead of using HTTP
headers, the framework now uses an URL pathname suffix, including outlet
name and remote or standard RSC rendering and static generation uses the
same format in filenames. This fixes #95.

Adds build options to enable compression (GZip and Brotli) and opt-out
of static RSC rendering, see more about this in the updated docs.

Adds static export option to render a single outlet. This is mainly to
keep consistency with the URL format right now, but might be a useful
feature for SSG.

Updates the deployment Adapter API to not copy compressed files by
default, but copy static RSC files and also the Vercel adapter to add
the necessary
`Content-Type` header for static `.x-component` files. This fixes #94.

Removes the `standalone` route option and mode as it was not really
useful and there are other ways to achieve the same functionality.