fix: handling http response headers (#106)
Moves route matching and route component loading into middleware when
using the file-system based router to remove async implementation in the
React Server Component entrypoint of the router.

Changes RSC payload handling to use the same approach as in the HTML SSR
renderer to handle async components better. Introduces a rendering lock
mechanism to suspend sending the response to be able to handle response
headers in async components until the lock is released.

Refactors HTTP response header handling to use `Headers` directly and
introduces helper functions to set, append, delete, or clear response
headers in the HTTP context.

Fixes an issue with accessing the `Response` sent to the client when
using `useResponse` in an async React Server Component.

Adds test cases for above changes and updates the docs to include the
new helpers and to give more details about how to handle response
headers.

#100