feat: support Bun runtime (#112)
This PR adds Bun support with a minimum version of Bun v1.1.45.

The most prominent change is the compatibility fallback workaround on
how streams are being used with the SSR worker. Node.js supports passing
streams to a worker, while Bun doesn't and it was necessary to pass
buffers directly instead of using a stream.

Adds workaround to switch `NODE_ENV` to production when using
`@lazarv/react-server` in production mode as React needs to resolve to
the appropriate module in production.

A Bun plugin was implemented to resolve the `react` and
`react/jsx-dev-runtime` modules to their `react-server` versions in
development mode.

Using `fsevents` in file watchers is disabled when using Bun, while the
server restart option is disabled in the development server when using
Bun as there's a known issue with Vite restarting the server.

Fixes how the logger instance is initiated in production mode.

Updates the "Get Started" page in the documentation to include
instructions on how to use Bun when working with `@lazarv/react-server`
and also includes a minor tweak on the styling of the documentation
sidebar navigation.