fix: handle server function not found (#205)
This PR changes how missing server functions are handled. Instead of
redirecting, a missing server function throws an error, which can be
captured using a `useActionState` during server-side rendering and can
also be captured using an error component or an error boundary to pass
control to the developer on how to handle the issue.