feat: rolldown-vite (#215)
Introducing `rolldown-vite` in `@lazarv/react-server`!

Makes every existing feature compatible with `rolldown-vite`.
The default Vite React plugin is now `@vitejs/plugin-react-oxc`. All
previous usage of `@vitejs/plugin-react-swc` was removed in favor of the
default plugin based on Oxc.

`RemoteComponent` now retrieves the browser manifest of the remote and
creates an import map to load the shared dependencies from the host.
Shared dependencies are now more automated as a result of the manual
handling of an import map being almost impossible, since hashes are no
longer based on file content.

Upgraded default Node.js version to 22.15. The minimum Bun version with
`rolldown-vite` is now set to 1.2.9.