feat: native formdata parser (#247)
Drops `parseMultipartFormData` and instead uses `request.formData()`.
Adds a new example for file uploads using a server component with
`useActionState` from `@lazarv/react-server/router` and a client
component using `useActionState` from `react`.
The example also includes a middleware for a minimal request body limit
checking demo.
Changes how middleware errors are handled and passed to the rendering
phase to be able to control what to render in case of an error thrown
when running middleware.