feat: use vite env loader (#258)
This PR enables Vite environment loading when using the development
server or in production builds, using the standard solution that Vite
provides, as described at
https://vite.dev/config/#using-environment-variables-in-config.

A new `--mode` option was introduced for the development server and
build CLI commands, which acts the same way as for Vite.

Vite configuration as a function is now supported the same way as in
Vite, with the only caveat that `isSsrBuild` and `isPreview` are not
implemented, as these are not useful when using the framework.

The development server also uses the configuration watcher to include
`.env` files and restarts the server on changes.

Minor fixes were also included regarding the Vite configuration merging.