fix: accept header handling (#279)
This PR fixes the render context flag for `isHTML` by using an HTML
response when the `Accept` header contains `text/html`, the header is
empty, or when the header is `*/*`. #277

Changes the CI workflow to run tests in parallel.
Fixes client module handling on Windows and fixes throwing a "Page not
found" error on non-page requests in the file-system based router
middleware chain.